+++ romancal/datamodels/__init__.py
@@ -1,15 +1,15 @@
 # Base Roman data model
-from .models.core import RomanDataModel
+from .core import RomanDataModel
 
 # Reference File data models
-from .models.reference_files.referencefile import ReferenceFileModel
-from .models.reference_files.flat import FlatModel
+from .reference_files.referencefile import ReferenceFileModel
+from .reference_files.flat import FlatModel
 
 # Level 1 data model
-from .models.level1 import Level1FileModel
+from .level1 import Level1FileModel
 
 # Image Models
-from .models.image import ImageModel
+from .image import ImageModel
 
 from .open_impl import open
