+++ romancal/datamodels/image.py
